Reorg of files, to make clear which files directly patch into the game.  Update mod version.  Add list of mods used in logging.
