+++ Source/Patches/DistrictManagerReleaseDistrictPatch.cs
@@ -1,0 +1,20 @@
+﻿using Harmony;
+
+namespace EnhancedDistrictServices
+{
+    [HarmonyPatch(typeof(DistrictManager))]
+    [HarmonyPatch("ReleaseDistrictImplementation")]
+    public class DistrictManagerReleaseDistrictPatch
+    {
+        public static bool Prefix(byte district, ref District data)
+        {
+            var districtPark = DistrictPark.FromDistrict(district);
+            if (districtPark.Name != string.Empty)
+            {
+                Constraints.ReleaseDistrictPark(districtPark);
+            }
+
+            return true;
+        }
+    }
+}
